Update KUDO Kafka deployment instructions
KUDO instance naming is now deterministic, which means that users no
longer need to specify instance names and we can simplify the deployment
instructions and omit the connection string for ZooKeeper.

This change still demonstrates an upgrade to the KUDO Kafka Operator as
well as Kafka itself by going from 2.3.0 / 1.0.0 to 2.3.1 / 1.0.1.